requirements.txt
selenium==4.25.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
rich==13.7.1
tabulate==0.9.0
Flask==3.0.3
pytest==8.3.3
